+++ packages/medusa-1c/src/modules/1c/models/onec.ts
@@ -1,0 +1,9 @@
+import { model } from "@medusajs/utils"
+
+export const Onec = model.define("onec", {
+	id: model.id().primaryKey(),
+	interval: model.number().nullable(),
+	chunkSize: model.number().nullable(),
+	useZip: model.boolean().default(false),
+	attributes: model.json().nullable(),
+})