+++ packages/site/src/components/pages/main/index.ts
@@ -1,3 +1,0 @@
-export { default as AnimatedLogos } from './AnimatedLogos';
-export { default as Links } from './Links';
-export { default as Hero } from './Hero';