+++ packages/site/src/components/pages/medusa/common/index.ts
@@ -1,6 +1,0 @@
-export { default as Filter } from "./Filter";
-export { default as Hero } from "./Hero";
-export { default as Pagination } from "./Pagination";
-export { default as Search } from "./Search";
-export { default as SortRow } from "./SortRow";
-export { default as SortSelect } from "./SortSelect";