+++ packages/site/src/components/pages/medusa/plugins-detail/index.ts
@@ -1,4 +1,0 @@
-export { default as PluginInfo } from "./PluginInfo";
-export { default as PluginInfoSupportBanner } from "./PluginInfoSupportBanner";
-export { default as PluginSidebar } from "./PluginSidebar";
-export { default as RelatedPlugins } from "./RelatedPlugins";