+++ packages/site/src/components/pages/medusa/plugins/index.ts
@@ -1,3 +1,0 @@
-export { default as PluginCard } from "./PluginCard";
-export { default as PluginsClientPage } from "./PluginsClientPage";
-export { default as PluginsList } from "./PluginsList";