+++ packages/medusa-feed-yandex/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @gorgo/medusa-feed-yandex
 
+## 0.2.28
+
+### Patch Changes
+
+- [#265](https://github.com/gorgojs/medusa-plugins/pull/265) [`d2cd44a`](https://github.com/gorgojs/medusa-plugins/commit/d2cd44a3fbe4fa1db50b6c63a9bc4683a48de459) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.15.3 and update readme
+
+## 0.2.27
+
+### Patch Changes
+
+- [#253](https://github.com/gorgojs/medusa-plugins/pull/253) [`98e79cd`](https://github.com/gorgojs/medusa-plugins/commit/98e79cd66117c0be832c4cafa265d2653d1754cc) Thanks [@ttokyose](https://github.com/ttokyose)! - add plugins logo
+
+## 0.2.26
+
+### Patch Changes
+
+- [`6586436`](https://github.com/gorgojs/medusa-plugins/commit/65864366a1e75568cd70f4674e73789f30f527fa) Thanks [@gorgohead](https://github.com/gorgohead)! - align telemetry
+
+## 0.2.25
+
+### Patch Changes
+
+- [#248](https://github.com/gorgojs/medusa-plugins/pull/248) [`44351ab`](https://github.com/gorgojs/medusa-plugins/commit/44351abcafb01ffa8f2c9e08fc98f3c359444b8d) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.15.2 and update readme
+
+- [#244](https://github.com/gorgojs/medusa-plugins/pull/244) [`eefd1db`](https://github.com/gorgojs/medusa-plugins/commit/eefd1dbb0762f0e6bb9b5916386a26ef8d3a266c) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.15.1 and update readme
+
+## 0.2.24
+
+### Patch Changes
+
+- [`277466a`](https://github.com/gorgojs/medusa-plugins/commit/277466a7b2bafe155cf5f7f5f2ef139d634c68fa) Thanks [@ttokyose](https://github.com/ttokyose)! - fix internal server error and improve docs
+
+- [#239](https://github.com/gorgojs/medusa-plugins/pull/239) [`bcd8eaf`](https://github.com/gorgojs/medusa-plugins/commit/bcd8eafeed495dbb752840322fafe9be84a8fd57) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.14.2 and update readme
+
+## 0.2.23
+
+### Patch Changes
+
+- [#230](https://github.com/gorgojs/medusa-plugins/pull/230) [`433f939`](https://github.com/gorgojs/medusa-plugins/commit/433f9399809b19b43ac5386a41674f78e99db264) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.14.1 and update readme
+
+## 0.2.22
+
+### Patch Changes
+
+- [#212](https://github.com/gorgojs/medusa-plugins/pull/212) [`7232152`](https://github.com/gorgojs/medusa-plugins/commit/7232152af5fcdcf57bc6ca53ade54327e55ee605) Thanks [@gorgohead](https://github.com/gorgohead)! - improve readme
+
+## 0.2.21
+
+### Patch Changes
+
+- [#198](https://github.com/gorgojs/medusa-plugins/pull/198) [`39cbe47`](https://github.com/gorgojs/medusa-plugins/commit/39cbe4766b87fa64d538961ac614e00679e45bf9) Thanks [@gorgohead](https://github.com/gorgohead)! - improve readme and docs
+
+## 0.2.20
+
+### Patch Changes
+
+- Test on Medusa 2.13.4
+
 ## 0.2.19
 
 ### Patch Changes