+++ packages/medusa-fulfillment-apiship/CHANGELOG.md
@@ -1,5 +1,97 @@
 # @gorgo/medusa-fulfillment-apiship
 
+## 0.5.6
+
+### Patch Changes
+
+- [#265](https://github.com/gorgojs/medusa-plugins/pull/265) [`d2cd44a`](https://github.com/gorgojs/medusa-plugins/commit/d2cd44a3fbe4fa1db50b6c63a9bc4683a48de459) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.15.3 and update readme
+
+## 0.5.5
+
+### Patch Changes
+
+- [#257](https://github.com/gorgojs/medusa-plugins/pull/257) [`3e9c650`](https://github.com/gorgojs/medusa-plugins/commit/3e9c6503eeaf937835b6ca67e0d40a8c24af8f3d) Thanks [@ttokyose](https://github.com/ttokyose)! - make the connection name optional
+
+## 0.5.4
+
+### Patch Changes
+
+- [#253](https://github.com/gorgojs/medusa-plugins/pull/253) [`98e79cd`](https://github.com/gorgojs/medusa-plugins/commit/98e79cd66117c0be832c4cafa265d2653d1754cc) Thanks [@ttokyose](https://github.com/ttokyose)! - add plugins logo
+
+## 0.5.3
+
+### Patch Changes
+
+- [`6586436`](https://github.com/gorgojs/medusa-plugins/commit/65864366a1e75568cd70f4674e73789f30f527fa) Thanks [@gorgohead](https://github.com/gorgohead)! - align telemetry
+
+## 0.5.2
+
+### Patch Changes
+
+- [#248](https://github.com/gorgojs/medusa-plugins/pull/248) [`44351ab`](https://github.com/gorgojs/medusa-plugins/commit/44351abcafb01ffa8f2c9e08fc98f3c359444b8d) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.15.2 and update readme
+
+- [#244](https://github.com/gorgojs/medusa-plugins/pull/244) [`eefd1db`](https://github.com/gorgojs/medusa-plugins/commit/eefd1dbb0762f0e6bb9b5916386a26ef8d3a266c) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.15.1 and update readme
+
+## 0.5.1
+
+### Patch Changes
+
+- [#239](https://github.com/gorgojs/medusa-plugins/pull/239) [`bcd8eaf`](https://github.com/gorgojs/medusa-plugins/commit/bcd8eafeed495dbb752840322fafe9be84a8fd57) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.14.2 and update readme
+
+## 0.5.0
+
+### Minor Changes
+
+- [#236](https://github.com/gorgojs/medusa-plugins/pull/236) [`a58f4e5`](https://github.com/gorgojs/medusa-plugins/commit/a58f4e5b57fcc450558c622fe823065aaca1dead) Thanks [@gorgohead](https://github.com/gorgohead)! - improve naming and requests
+
+## 0.4.4
+
+### Patch Changes
+
+- [#230](https://github.com/gorgojs/medusa-plugins/pull/230) [`433f939`](https://github.com/gorgojs/medusa-plugins/commit/433f9399809b19b43ac5386a41674f78e99db264) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.14.1 and update readme
+
+## 0.4.3
+
+### Patch Changes
+
+- [#212](https://github.com/gorgojs/medusa-plugins/pull/212) [`7232152`](https://github.com/gorgojs/medusa-plugins/commit/7232152af5fcdcf57bc6ca53ade54327e55ee605) Thanks [@gorgohead](https://github.com/gorgohead)! - improve readme
+
+## 0.4.2
+
+### Patch Changes
+
+- [#196](https://github.com/gorgojs/medusa-plugins/pull/196) [`0bd3966`](https://github.com/gorgojs/medusa-plugins/commit/0bd39665e47946f2d8c0968d27b8ff92064113d2) Thanks [@gorgohead](https://github.com/gorgohead)! - improve readme
+
+## 0.4.1
+
+### Patch Changes
+
+- Update docs
+
+## 0.4.0
+
+### Minor Changes
+
+- [#190](https://github.com/gorgojs/medusa-plugins/pull/190) [`eab0baf`](https://github.com/gorgojs/medusa-plugins/commit/eab0baf75e03aedc86df178ee354403058a30547) Thanks [@gorgohead](https://github.com/gorgohead)! - add apiship readme demo video ([#190](https://github.com/gorgojs/medusa-plugins/issues/190))
+
+## 0.3.1
+
+### Patch Changes
+
+- [`deef5f7`](https://github.com/gorgojs/medusa-plugins/commit/deef5f72f36a835bbf766dbf3b6b5dcece9db4a8) Thanks [@gorgohead](https://github.com/gorgohead)! - improve UI texts
+
+## 0.3.0
+
+### Minor Changes
+
+- [#182](https://github.com/gorgojs/medusa-plugins/pull/182) [`8463123`](https://github.com/gorgojs/medusa-plugins/commit/84631238073bf5803cab1f7c29a1681c1ab496db) Thanks [@gorgohead](https://github.com/gorgohead)! - feat: manage all ApiShip settings in Medusa Admin
+
+## 0.2.4
+
+### Patch Changes
+
+- Test on Medusa 2.13.4
+
 ## 0.2.3
 
 ### Patch Changes