+++ packages/medusa-payment-tkassa/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @gorgo/medusa-payment-tkassa
 
+## 0.4.3
+
+### Patch Changes
+
+- [#265](https://github.com/gorgojs/medusa-plugins/pull/265) [`2ef7450`](https://github.com/gorgojs/medusa-plugins/commit/2ef74506d5321c30d7b3f44cbd8c7cac442cb9ee) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - add integration test script
+
+- [#265](https://github.com/gorgojs/medusa-plugins/pull/265) [`0940641`](https://github.com/gorgojs/medusa-plugins/commit/09406410682d438050c640982c4c46e6d496b6cb) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - update docs
+
+- [#265](https://github.com/gorgojs/medusa-plugins/pull/265) [`e47a625`](https://github.com/gorgojs/medusa-plugins/commit/e47a6257c8a5212a4ceee0657c60ab4574ece302) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - add unit tests for core provider methods and utils
+
+- [#265](https://github.com/gorgojs/medusa-plugins/pull/265) [`6a43601`](https://github.com/gorgojs/medusa-plugins/commit/6a43601776f04edae98c391d583baa5d53d75b14) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - fix dependencies
+
+- [#265](https://github.com/gorgojs/medusa-plugins/pull/265) [`d2cd44a`](https://github.com/gorgojs/medusa-plugins/commit/d2cd44a3fbe4fa1db50b6c63a9bc4683a48de459) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.15.3 and update readme
+
+## 0.4.2
+
+### Patch Changes
+
+- [#253](https://github.com/gorgojs/medusa-plugins/pull/253) [`98e79cd`](https://github.com/gorgojs/medusa-plugins/commit/98e79cd66117c0be832c4cafa265d2653d1754cc) Thanks [@ttokyose](https://github.com/ttokyose)! - add plugins logo
+
+## 0.4.1
+
+### Patch Changes
+
+- [`6586436`](https://github.com/gorgojs/medusa-plugins/commit/65864366a1e75568cd70f4674e73789f30f527fa) Thanks [@gorgohead](https://github.com/gorgohead)! - align telemetry
+
+## 0.4.0
+
+### Minor Changes
+
+- [#249](https://github.com/gorgojs/medusa-plugins/pull/249) [`7f4fca9`](https://github.com/gorgojs/medusa-plugins/commit/7f4fca92882ec00ae3c17d3e61eaee742a073f64) Thanks [@gorgohead](https://github.com/gorgohead)! - add telemetry
+
+## 0.3.22
+
+### Patch Changes
+
+- [#248](https://github.com/gorgojs/medusa-plugins/pull/248) [`44351ab`](https://github.com/gorgojs/medusa-plugins/commit/44351abcafb01ffa8f2c9e08fc98f3c359444b8d) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.15.2 and update readme
+
+- [`d919655`](https://github.com/gorgojs/medusa-plugins/commit/d91965567a662d735f2d122058d5c9b9fe2392b9) Thanks [@gorgohead](https://github.com/gorgohead)! - add used by t-kassa section
+
+- [#244](https://github.com/gorgojs/medusa-plugins/pull/244) [`eefd1db`](https://github.com/gorgojs/medusa-plugins/commit/eefd1dbb0762f0e6bb9b5916386a26ef8d3a266c) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.15.1 and update readme
+
+## 0.3.21
+
+### Patch Changes
+
+- [#239](https://github.com/gorgojs/medusa-plugins/pull/239) [`bcd8eaf`](https://github.com/gorgojs/medusa-plugins/commit/bcd8eafeed495dbb752840322fafe9be84a8fd57) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.14.2 and update readme
+
+## 0.3.20
+
+### Patch Changes
+
+- [#230](https://github.com/gorgojs/medusa-plugins/pull/230) [`433f939`](https://github.com/gorgojs/medusa-plugins/commit/433f9399809b19b43ac5386a41674f78e99db264) Thanks [@gorgo-app](https://github.com/apps/gorgo-app)! - test on Medusa v2.14.1 and update readme
+
+## 0.3.19
+
+### Patch Changes
+
+- [#212](https://github.com/gorgojs/medusa-plugins/pull/212) [`7232152`](https://github.com/gorgojs/medusa-plugins/commit/7232152af5fcdcf57bc6ca53ade54327e55ee605) Thanks [@gorgohead](https://github.com/gorgohead)! - improve readme
+
+## 0.3.18
+
+### Patch Changes
+
+- [#198](https://github.com/gorgojs/medusa-plugins/pull/198) [`39cbe47`](https://github.com/gorgojs/medusa-plugins/commit/39cbe4766b87fa64d538961ac614e00679e45bf9) Thanks [@gorgohead](https://github.com/gorgohead)! - improve readme and docs
+
+## 0.3.17
+
+### Patch Changes
+
+- Test on Medusa 2.13.4
+
 ## 0.3.16
 
 ### Patch Changes